requirements.txt
reflex>=0.6.4
firebase_admin
python-dotenv
reflex-nav-menu
reflex-calendar